refactor sequence component into smaller ones
